requirements.txt
boto3==1.35.8
python-dotenv==1.0.1
